+++ doc/changes.rst
@@ -5,6 +5,8 @@ desitarget Change Log
 0.15.0 (unreleased)
 -------------------
 
+* Refactored :mod:`desitarget.QA` to calculate density fluctuations in HEALPixels 
+  instead of in bricks [`PR #217`_]:
 * Updated :mod:`desitarget.io` for the DR5 RELEASE number [`PR #214`_]:
 * Updated :mod:`desitarget.QA` to produce QA plots [`PR #210`_]:
    * Has a simple binary that runs the plot-making software in full
@@ -19,11 +21,11 @@ desitarget Change Log
 * Added ``mpi_select_mock_targets``
 * Added ``desitarget.mock.sky.random_sky`` [`PR #219`_]
 
-
 .. _`PR #200`: https://github.com/desihub/desitarget/pull/200
 .. _`PR #210`: https://github.com/desihub/desitarget/pull/210
 .. _`PR #214`: https://github.com/desihub/desitarget/pull/214
 .. _`PR #219`: https://github.com/desihub/desitarget/pull/219
+.. _`PR #217`: https://github.com/desihub/desitarget/pull/217
 
 0.14.0 (2017-07-10)
 -------------------